+++ nanodjango/settings.py
@@ -33,11 +33,13 @@ INSTALLED_APPS = [
     "django.contrib.contenttypes",
     "django.contrib.sessions",
     "django.contrib.messages",
+    "whitenoise.runserver_nostatic",
     "django.contrib.staticfiles",
 ] + app_conf.get("EXTRA_APPS", [])
 
 MIDDLEWARE = [
     "django.middleware.security.SecurityMiddleware",
+    "whitenoise.middleware.WhiteNoiseMiddleware",
     "django.contrib.sessions.middleware.SessionMiddleware",
     "django.middleware.common.CommonMiddleware",
     "django.middleware.csrf.CsrfViewMiddleware",
@@ -96,12 +98,35 @@ STATIC_URL = "/static/"
 STATICFILES_DIRS = [
     BASE_DIR / "static",
 ]
+STATIC_ROOT = BASE_DIR / "static-collected"
 
 # Media files
 MEDIA_URL = "/media/"
 MEDIA_ROOT = BASE_DIR / "media"
 
+STORAGES = {
+    # Django default:
+    "default": {
+        "BACKEND": "django.core.files.storage.FileSystemStorage",
+    },
+    # Whitenoise:
+    "staticfiles": {
+        "BACKEND": "whitenoise.storage.CompressedManifestStaticFilesStorage",
+    },
+}
+
 
 # nanodjango specific config
+
+# URL to serve admin site on
+#
+# If set, admin site will always be registered
+#
+# If not set, defaults to "/admin/" when the @app.admin decorator is used
 ADMIN_URL = None
+
+# URL to serve the API on, if @app.api is used
 API_URL = "api/"
+
+# Directory containing public files
+PUBLIC_DIR = BASE_DIR / "public"